Completo

+++ ML/Cine.cs
@@ -14,8 +14,8 @@ namespace ML
         public decimal Venta { get; set; }
         public ML.Zona Zona { get; set; }
         public List<Object> Cines { get; set; } 
-        public decimal Latitud { get; set; }
-        public decimal Longitud { get; set; }
+        public double Latitud { get; set; }
+        public double Longitud { get; set; }
         public ML.Estadistica Estadistica { get; set; }
     }
 }
